+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `StaticDependenciesProperty`
 - Schema validation via CLI
 - Allow restricting generation to (in)valid documents
+- Allow setting random seed
 
 ### Changed
 - Updated jsonoid-discovery dependency to v0.7.3